+++ src/main/java/com/minte9/threads/lock/SynchronizedApp.java
@@ -46,7 +46,7 @@ class SynchronizedRunner implements Runnable {
         }
     }
 
-    private void sleep(int miliseconds) {
+    private void sleep(long miliseconds) {
         try {
             Thread.sleep(miliseconds); // sleep - threads take turns
         } catch (InterruptedException e) {
